+++ dom/src/index.ts
@@ -1,5 +1,5 @@
 export * from './types';
-export * from './makeDomDriver';
+export * from './makeDomComponent';
 export * from './mockDomSource';
 export * from './api-wrappers';
 export * from 'mostly-dom';